Extract the endpoint property definitions to a new module, to address potential cyclic dependency issues
