+++ src/forges/forge.py
@@ -14,7 +14,8 @@ from src.blueprints.dataset import Dataset, DatasetParameterSet
 from src.blueprints.experiment import Experiment, ExperimentParameterSet
 from src.blueprints.project import Project, ProjectParameterSet
 from src.mytardis_client.data_types import URI, HttpRequestMethod
-from src.mytardis_client.endpoints import MyTardisEndpoint, get_endpoint_info
+from src.mytardis_client.endpoint_props import get_endpoint_info
+from src.mytardis_client.endpoints import MyTardisEndpoint
 from src.mytardis_client.mt_rest import BadGateWayException, MyTardisRESTFactory
 
 logger = logging.getLogger(__name__)
